requirements.txt
google-generativeai
google-adk
pandas
pymongo
python-dotenv
pdfplumber
tabula-py